Spearmint/ioquake3 3330 resync. Made by two heroes of the Olymp... Eugene C. and Zack Middleton. What more do we want?

+++ code/renderercommon/tr_common.h
@@ -116,8 +116,8 @@ extern	cvar_t	*r_fontForceAutoHint;
 
 qboolean	R_GetModeInfo( int *width, int *height, float *windowAspect, int mode );
 
-float R_NoiseGet4f( float x, float y, float z, float t );
-int R_RandomOn( float t );
+float R_NoiseGet4f( float x, float y, float z, double t );
+int R_RandomOn( double t );
 void  R_NoiseInit( void );
 
 image_t     *R_FindImageFile( const char *name, imgType_t type, imgFlags_t flags );
