+++ code/renderercommon/tr_noise.c
@@ -49,7 +49,7 @@ void R_NoiseInit( void )
 	}
 }
 
-float R_NoiseGet4f( float x, float y, float z, float t )
+float R_NoiseGet4f( float x, float y, float z, double t )
 {
 	int i;
 	int ix, iy, iz, it;
@@ -91,7 +91,7 @@ float R_NoiseGet4f( float x, float y, float z, float t )
 }
 
 // Used in the shader functions (GF_RANDOM) to implement a quasi random flickering.
-int R_RandomOn(float t)
+int R_RandomOn(double t)
 {
 	return VALR((unsigned int) floor(t));
 }
